some more changes
Standardized the username field on the API pages.
Changed the validateAPIKey() to set the local message variable and
return an array with the same message.
Added a return to the end of FindLanguageType().
Moved the language class to inside dbcore.
Moved the wdbmail class to inside dbcore.
Removed unneeded requires because of __autoload().
Changed some more api methods to use bound parameters.