lots of changes
Fixed cookies to set to http or https, depending on the host_url config
value.
Login and log out are working, along with the cookies that they set.
Permissions variable is set with the users permission level, but nothing
obeys it yet.
Fixed auth and encry missing value defaults in all.php
Moved LoginCheck() into the core, since everything is should be doing
that check.
Started finalizing the AP Fetch page.
Added separate link for Log out at the end of all the other navigation
links.