Some quick changes
Added Mac Address validation for each AP, If the values for a MAC
Address are not within the normal range for a MAC, the AP is dropped.
Added FindHighestSig function, to calculate the Highest Signal when one
is not in the VS1 file. Then will calculate the Highest RSSI off of that
value.
GPS date is now truely standardized by using "date($this->date_format,
strtotime($gps_line[10]))"
Added test script for FindHighestSig