safe keeping
working on moving the region generated data to a separate folder path
also made a threading launching script. right now it is hard coded to
launch 20 independent importd.php scripts at once.
be warned, this uses a lot of CPU and RAM, more so that there are 20
import daemons running.
Andrew ran a single import daemon and it took him roughly 9 hours to do
all the files in the database. I ran 20 import daemons and was able to
import all the files in roughly 1 hour and 40 minutes.