more changes
Added the Files_ID to the signals table so that the imports signal
history can be tracked.

Added a completed flag on the files table so we can tell what files have
crashed the import daemon, or not completed correctly.

Removed the files and files_tmp tables from the badcleanup function