Big Commit, Read the Notes
Changed the SWITCH_SCREEN constant. API is now part of SWITCH_SCREEN.
Each type of API is now defined as the SWITCH_EXTRAS to define the type
of API that is being used.

Fixed the KML Export Signal types and added signalPlotType to the GET
and POST for the export api.
1 = Points
2 = Signal Line
3 = 3D Signal Plot
4 = Both Signal Line and 3D Plot

Fixed the DATETIME column type in the blank_db.sql template

Fixed typo in the graph_results.tpl file.

Changed the default layout of the Index a little.

Changed the create cookies test from checking CLI and API to checking
HTML, makes it a little shorter code wise.

Added in autoload class paths for the RSS Feed writer.

That is about all I can think of for now.