Re-doing the Live AP Spaghetti code
It was a mess, I mean a complete mess of spaghetti code. Just wasn't
worth the effort to fork through it, so I am just going to boil it down
to the front end being reused (mostly) and re-pressing the backend code.

first up is redoing the table structure. moved user info into its own
table with the session_id.