Major Change to Live AP imports
Live AP imports now need to be initiated with username, apikey, title,
and notes post and will return a session id to be used when submitting
AP data.
Also still support the Old way of doing it, where a sessionID is
submitted with the AP data but no title and notes. This will use the
SessionID as the Title, since before it was not tracking APs by title
and just sessionid. Titles implementation gives more customizable
control over imports.

Also Updated the ValidateAPIKey() and fixed some bugs.

Moving Cookie Generation code to its own function. It does not seem to
be working right now, will need to work on it some more.