Two Fixes
Fixed the next run date time showing as NaN in firefox by changing the
value stored from date/time to unix_timestamp.

This required the daemons to be updated to check for an integer and the
daemon::SetNextJob and daemon::SetStartJob functions to set an integer
instead of a string.