Fixed Date/Time bug in Live AP Import
I had changed the date/time in the live AP API import page and that was
removing the micro-seconds from the data. Reverted back to the
micro-seconds being stored in the tables for Signals, GPS, and FA / LA
times.

Also updated the title management code to update the timestamp for the
titles, so that we can tell when titles need to be cleaned out.