Had to Change the PDO Error checking
Since PDO does not set the error number or error code after the first
query. But the execute and query functions will return FALSE on a failed
query. So instead of checking the error number for failure, I am now
just checking the execute and query results for failure. Apparently the
PHP devs do not consider it a bug... WTF...seriously?