Wubba lubba dub-dub!
Lots of changes,
working on the maps view for the Live AP view.
Attempt at proxy pass for websockets, but need to work on it.
need to work on shoehorning SSL support into the websockets library I am
using, since switching to another one would be too much of a pain in the
ass and i think shoehorning would be easier.
removed old scheduling code from new scheduling since its not needed,
still need to test it for errors.