Lots and lots of changes
Started work on the Federation Frontend interface.
This will use CURL to get the data from the API of the remote server and
then display it on the Frontend UI.
The /opt/federation.php switching is pretty much in place and how I
would like it, just need to make the backend fetch the actual data and
create the data layout smarty templates. create a few of them, but the
contents is just a copy of one of the templates, they will need some
work..... I hate UI design...