Finished #2 and a few bugs
Front end for the AP Efficiency is in place. That should close out issue
#2.

Fixed a bug where the Last Active and First Active date/time was not
being fetched correctly so the pointers table would have an empty field.

Also pretty sure I fixed the schedule status not being set back to
Waiting at the end of a run.