Testing maplibre 2.0.0 alpha. Added some IE compatibility
Testing maplibre 3d terrain which is not quite finished and a bit buggy still.  This is based on maplibre 2.0.0, which drops support for IE. since I didn't quite want to drop IE yet, I added a way to detect IE and use the older library and disable some incompatible features.